# br/storage: disable keep alive for GCS on default to speed up request